requirements.txt
# Your requirements go here
cachetools==4.2.1
google-auth==1.27.0
google-auth-oauthlib==0.4.2
gspread==3.7.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.0
rsa==4.7.2
colorama==0.4.6